requirements.txt
openai
rich
python-dotenv
click
readline; platform_system != "Windows" 